Extract service interfaces and types into contracts sub-package (PSY-84)
Create internal/services/contracts/ with all 31 service interfaces and
~120 shared types organized into domain files (catalog, auth, pipeline,
engagement, notification, admin, user, label, release, festival).
Original definitions replaced with type aliases for full backward
compatibility — no handler, route, or container changes needed.

Closes PSY-84
