+++ backend/internal/services/data_sync.go
@@ -29,80 +29,6 @@ func NewDataSyncService(database *gorm.DB) *DataSyncService {
 	}
 }
 
-// ExportedArtist represents an artist for export/import
-type ExportedArtist struct {
-	Name             string  `json:"name"`
-	City             *string `json:"city,omitempty"`
-	State            *string `json:"state,omitempty"`
-	BandcampEmbedURL *string `json:"bandcampEmbedUrl,omitempty"`
-	Instagram        *string `json:"instagram,omitempty"`
-	Facebook         *string `json:"facebook,omitempty"`
-	Twitter          *string `json:"twitter,omitempty"`
-	YouTube          *string `json:"youtube,omitempty"`
-	Spotify          *string `json:"spotify,omitempty"`
-	SoundCloud       *string `json:"soundcloud,omitempty"`
-	Bandcamp         *string `json:"bandcamp,omitempty"`
-	Website          *string `json:"website,omitempty"`
-}
-
-// ExportedVenue represents a venue for export/import
-type ExportedVenue struct {
-	Name       string  `json:"name"`
-	Address    *string `json:"address,omitempty"`
-	City       string  `json:"city"`
-	State      string  `json:"state"`
-	Zipcode    *string `json:"zipcode,omitempty"`
-	Verified   bool    `json:"verified"`
-	Instagram  *string `json:"instagram,omitempty"`
-	Facebook   *string `json:"facebook,omitempty"`
-	Twitter    *string `json:"twitter,omitempty"`
-	YouTube    *string `json:"youtube,omitempty"`
-	Spotify    *string `json:"spotify,omitempty"`
-	SoundCloud *string `json:"soundcloud,omitempty"`
-	Bandcamp   *string `json:"bandcamp,omitempty"`
-	Website    *string `json:"website,omitempty"`
-}
-
-// ExportedShowArtist represents an artist in a show lineup
-type ExportedShowArtist struct {
-	Name     string `json:"name"`
-	Position int    `json:"position"`
-	SetType  string `json:"setType"`
-}
-
-// ExportedShow represents a show for export/import
-type ExportedShow struct {
-	Title          string               `json:"title"`
-	EventDate      string               `json:"eventDate"` // ISO format
-	City           *string              `json:"city,omitempty"`
-	State          *string              `json:"state,omitempty"`
-	Price          *float64             `json:"price,omitempty"`
-	AgeRequirement *string              `json:"ageRequirement,omitempty"`
-	Description    *string              `json:"description,omitempty"`
-	Status         string               `json:"status"`
-	IsSoldOut      bool                 `json:"isSoldOut"`
-	IsCancelled    bool                 `json:"isCancelled"`
-	Venues         []ExportedVenue      `json:"venues"`
-	Artists        []ExportedShowArtist `json:"artists"`
-}
-
-// ExportShowsParams contains filters for show export
-type ExportShowsParams struct {
-	Limit      int
-	Offset     int
-	Status     string // "approved", "pending", "all"
-	FromDate   *time.Time
-	City       string
-	State      string
-	IncludeAll bool // Include all related data
-}
-
-// ExportShowsResult contains exported shows with pagination info
-type ExportShowsResult struct {
-	Shows []ExportedShow `json:"shows"`
-	Total int64          `json:"total"`
-}
-
 // ExportShows exports shows with their artists and venues
 func (s *DataSyncService) ExportShows(params ExportShowsParams) (*ExportShowsResult, error) {
 	if s.db == nil {
@@ -249,18 +175,6 @@ func (s *DataSyncService) ExportShows(params ExportShowsParams) (*ExportShowsRes
 }
 
 // ExportArtistsParams contains filters for artist export
-type ExportArtistsParams struct {
-	Limit  int
-	Offset int
-	Search string
-}
-
-// ExportArtistsResult contains exported artists with pagination info
-type ExportArtistsResult struct {
-	Artists []ExportedArtist `json:"artists"`
-	Total   int64            `json:"total"`
-}
-
 // ExportArtists exports artists
 func (s *DataSyncService) ExportArtists(params ExportArtistsParams) (*ExportArtistsResult, error) {
 	if s.db == nil {
@@ -319,21 +233,6 @@ func (s *DataSyncService) ExportArtists(params ExportArtistsParams) (*ExportArti
 }
 
 // ExportVenuesParams contains filters for venue export
-type ExportVenuesParams struct {
-	Limit    int
-	Offset   int
-	Search   string
-	Verified *bool
-	City     string
-	State    string
-}
-
-// ExportVenuesResult contains exported venues with pagination info
-type ExportVenuesResult struct {
-	Venues []ExportedVenue `json:"venues"`
-	Total  int64           `json:"total"`
-}
-
 // ExportVenues exports venues
 func (s *DataSyncService) ExportVenues(params ExportVenuesParams) (*ExportVenuesResult, error) {
 	if s.db == nil {
@@ -403,40 +302,6 @@ func (s *DataSyncService) ExportVenues(params ExportVenuesParams) (*ExportVenues
 }
 
 // DataImportRequest represents a batch import request
-type DataImportRequest struct {
-	Shows   []ExportedShow   `json:"shows,omitempty"`
-	Artists []ExportedArtist `json:"artists,omitempty"`
-	Venues  []ExportedVenue  `json:"venues,omitempty"`
-	DryRun  bool             `json:"dryRun"`
-}
-
-// DataImportResult contains statistics about the import operation
-type DataImportResult struct {
-	Shows struct {
-		Total      int      `json:"total"`
-		Imported   int      `json:"imported"`
-		Duplicates int      `json:"duplicates"`
-		Errors     int      `json:"errors"`
-		Messages   []string `json:"messages"`
-	} `json:"shows"`
-	Artists struct {
-		Total      int      `json:"total"`
-		Imported   int      `json:"imported"`
-		Duplicates int      `json:"duplicates"`
-		Updated    int      `json:"updated"`
-		Errors     int      `json:"errors"`
-		Messages   []string `json:"messages"`
-	} `json:"artists"`
-	Venues struct {
-		Total      int      `json:"total"`
-		Imported   int      `json:"imported"`
-		Duplicates int      `json:"duplicates"`
-		Updated    int      `json:"updated"`
-		Errors     int      `json:"errors"`
-		Messages   []string `json:"messages"`
-	} `json:"venues"`
-}
-
 // ImportData imports shows, artists, and venues with deduplication
 func (s *DataSyncService) ImportData(req DataImportRequest) (*DataImportResult, error) {
 	if s.db == nil {
