+++ backend/internal/services/extraction.go
@@ -58,71 +58,6 @@ func NewExtractionService(database *gorm.DB, cfg *config.Config) *ExtractionServ
 	}
 }
 
-// ExtractShowRequest represents the extraction request
-type ExtractShowRequest struct {
-	Type      string `json:"type"`       // "text", "image", or "both"
-	Text      string `json:"text"`       // Text content
-	ImageData string `json:"image_data"` // Base64-encoded image
-	MediaType string `json:"media_type"` // MIME type of image
-}
-
-// MatchSuggestion represents a close-but-not-exact match from the database
-type MatchSuggestion struct {
-	ID   uint   `json:"id"`
-	Name string `json:"name"`
-	Slug string `json:"slug"`
-}
-
-// VenueMatchSuggestion includes location info
-type VenueMatchSuggestion struct {
-	ID    uint   `json:"id"`
-	Name  string `json:"name"`
-	Slug  string `json:"slug"`
-	City  string `json:"city"`
-	State string `json:"state"`
-}
-
-// ExtractedArtist represents an extracted artist with optional DB match
-type ExtractedArtist struct {
-	Name            string            `json:"name"`
-	IsHeadliner     bool              `json:"is_headliner"`
-	InstagramHandle string            `json:"instagram_handle,omitempty"`
-	MatchedID       *uint             `json:"matched_id,omitempty"`
-	MatchedName     *string           `json:"matched_name,omitempty"`
-	MatchedSlug     *string           `json:"matched_slug,omitempty"`
-	Suggestions     []MatchSuggestion `json:"suggestions,omitempty"`
-}
-
-// ExtractedVenue represents an extracted venue with optional DB match
-type ExtractedVenue struct {
-	Name        string                 `json:"name"`
-	City        string                 `json:"city,omitempty"`
-	State       string                 `json:"state,omitempty"`
-	MatchedID   *uint                  `json:"matched_id,omitempty"`
-	MatchedName *string                `json:"matched_name,omitempty"`
-	MatchedSlug *string                `json:"matched_slug,omitempty"`
-	Suggestions []VenueMatchSuggestion `json:"suggestions,omitempty"`
-}
-
-// ExtractedShowData is the full extraction result
-type ExtractedShowData struct {
-	Artists     []ExtractedArtist `json:"artists"`
-	Venue       *ExtractedVenue   `json:"venue,omitempty"`
-	Date        string            `json:"date,omitempty"`
-	Time        string            `json:"time,omitempty"`
-	Cost        string            `json:"cost,omitempty"`
-	Ages        string            `json:"ages,omitempty"`
-	Description string            `json:"description,omitempty"`
-}
-
-// ExtractShowResponse is the API response wrapper
-type ExtractShowResponse struct {
-	Success  bool               `json:"success"`
-	Data     *ExtractedShowData `json:"data,omitempty"`
-	Error    string             `json:"error,omitempty"`
-	Warnings []string           `json:"warnings,omitempty"`
-}
-
 // ExtractShow processes text or image input through Claude and matches against the database
 func (s *ExtractionService) ExtractShow(req *ExtractShowRequest) (*ExtractShowResponse, error) {
 	if s.config.Anthropic.APIKey == "" {
